tests/submission/PhilipTaylor/canvas/2d.text.align.valid.html

<!DOCTYPE html>
<html><head><meta charset="utf-8">
<title>Canvas test: 2d.text.align.valid</title>
<meta content="Philip Taylor" name="author">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="../common/canvas-tests.js"></script>
<link href="../common/canvas-tests.css" rel="stylesheet">
</head><body class="show_output">

<h1><a href="index.2d.html">2d</a>.<a href="index.2d.text.html">text</a>.<a href="index.2d.text.align.html">align</a>.valid</h1>
<p class="desc"></p>

<div class="refs">Spec references:
<ul>
<li><a href="../annotated-spec/canvas.html#testrefs.2d.text.align.get">2d.text.align.get</a>
</li><li><a href="../annotated-spec/canvas.html#testrefs.2d.text.align.set">2d.text.align.set</a>
</li></ul>
</div>

<p class="output">Actual output:</p>
<canvas class="output" height="50" id="c" width="100"><p class="fallback">FAIL (fallback content)</p></canvas>

<ul id="d"></ul>
<script>
var t = async_test("");
_addTest(function(canvas, ctx) {

ctx.textAlign = 'start';
_assertSame(ctx.textAlign, 'start', "ctx.textAlign", "'start'");

ctx.textAlign = 'end';
_assertSame(ctx.textAlign, 'end', "ctx.textAlign", "'end'");

ctx.textAlign = 'left';
_assertSame(ctx.textAlign, 'left', "ctx.textAlign", "'left'");

ctx.textAlign = 'right';
_assertSame(ctx.textAlign, 'right', "ctx.textAlign", "'right'");

ctx.textAlign = 'center';
_assertSame(ctx.textAlign, 'center', "ctx.textAlign", "'center'");


});
</script>

<div id="log"></div>
</body></html>
